requirements.txt
pinecone-client
openai
langchain
python-dotenv
pypdf
tiktoken